+++ core/src/main/java/com/trendyol/design/core/badge/boxbadge/model/BoxBadgeGroupContent.kt
@@ -4,6 +4,6 @@ import androidx.compose.runtime.Stable
 import com.trendyol.design.core.badge.boxbadge.BoxBadgeType
 
 @Stable
-data class BoxBadgeListContent(
+data class BoxBadgeGroupContent(
     val list: List<BoxBadgeType>
 )